+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Add `throttle` method decorator
 * Add `Blast.requestIdleCallback` implementation
 * Removed `romanize` string methods
+* Removed `Function.curry` function
 
 ## 0.6.1 (2018-10-16)